Rename Atom Lite namespace and projects
