+++ Tests/AtomLiteTestApp/Properties/AssemblyInfo.cs
