+++ nanoFramework.AtomLite.nuspec
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
   <metadata>
-    <id>nanoFramework.M5AtomLite</id>
+    <id>nanoFramework.AtomLite</id>
     <version>$version$</version>
-    <title>nanoFramework.M5AtomLite</title>
+    <title>nanoFramework.AtomLite</title>
     <authors>nanoFramework project contributors</authors>
     <owners>nanoFramework,dotnetfoundation</owners>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
@@ -16,8 +16,8 @@
     <icon>images\nf-logo.png</icon>
     <repository type="git" url="https://github.com/nanoframework/nanoFramework.M5Stack" commit="$commit$" />
     <copyright>Copyright (c) .NET Foundation and Contributors</copyright>
-    <description>This package includes the nanoFramework.M5AtomLite assembly for .NET nanoFramework C# projects.</description>
-    <tags>nanoFramework C# csharp netmf netnf nanoFramework.M5AtomLite</tags>
+    <description>This package includes the nanoFramework.AtomLite assembly for .NET nanoFramework C# projects.</description>
+    <tags>nanoFramework C# csharp netmf netnf nanoFramework.AtomLite</tags>
     <dependencies>
       <dependency id="nanoFramework.CoreLibrary" version="1.12.0-preview.5" />
       <dependency id="nanoFramework.Hardware.Esp32" version="1.3.5-preview.6" />
@@ -33,11 +33,11 @@
     </dependencies>
   </metadata>
   <files>
-    <file src="nanoFramework.M5AtomLite\bin\Release\nanoFramework.M5AtomLite.dll" target="lib\nanoFramework.M5AtomLite.dll" />
-    <file src="nanoFramework.M5AtomLite\bin\Release\nanoFramework.M5AtomLite.pdb" target="lib\nanoFramework.M5AtomLite.pdb" />
-    <file src="nanoFramework.M5AtomLite\bin\Release\nanoFramework.M5AtomLite.pdbx" target="lib\nanoFramework.M5AtomLite.pdbx" />
-    <file src="nanoFramework.M5AtomLite\bin\Release\nanoFramework.M5AtomLite.pe" target="lib\nanoFramework.M5AtomLite.pe" />
-    <file src="nanoFramework.M5AtomLite\bin\Release\nanoFramework.M5AtomLite.xml" target="lib\nanoFramework.M5AtomLite.xml" />
+    <file src="nanoFramework.AtomLite\bin\Release\nanoFramework.AtomLite.dll" target="lib\nanoFramework.AtomLite.dll" />
+    <file src="nanoFramework.AtomLite\bin\Release\nanoFramework.AtomLite.pdb" target="lib\nanoFramework.AtomLite.pdb" />
+    <file src="nanoFramework.AtomLite\bin\Release\nanoFramework.AtomLite.pdbx" target="lib\nanoFramework.AtomLite.pdbx" />
+    <file src="nanoFramework.AtomLite\bin\Release\nanoFramework.AtomLite.pe" target="lib\nanoFramework.AtomLite.pe" />
+    <file src="nanoFramework.AtomLite\bin\Release\nanoFramework.AtomLite.xml" target="lib\nanoFramework.AtomLite.xml" />
     <file src="assets\readme.txt" target="" />
     <file src="README.md" target="docs\" />
     <file src="assets\nf-logo.png" target="images\" />
