+++ nanoFramework.AtomLite/Properties/AssemblyInfo.cs
