+++ nanoFramework.AtomLite/key.snk
