+++ nanoFramework.AtomLite/packages.config
