+++ nanoFramework.M5Stack.sln
@@ -6,7 +6,7 @@ Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.M5Core", "nan
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{0A689EB3-86E0-448E-99F4-3D644EC3D8C5}"
 	ProjectSection(SolutionItems) = preProject
-		nanoFramework.M5AtomLite.nuspec = nanoFramework.M5AtomLite.nuspec
+		nanoFramework.AtomLite.nuspec = nanoFramework.AtomLite.nuspec
 		nanoFramework.M5Core.nuspec = nanoFramework.M5Core.nuspec
 		nanoFramework.M5Core2.nuspec = nanoFramework.M5Core2.nuspec
 		nanoFramework.M5StickC.nuspec = nanoFramework.M5StickC.nuspec
@@ -37,9 +37,9 @@ Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "nanoFramework.M5StackCore",
 EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "M5Core2TestApp", "Tests\M5Core2TestApp\M5Core2TestApp.nfproj", "{20266750-53F3-46D5-8626-1438AC985033}"
 EndProject
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.M5AtomLite", "nanoFramework.M5AtomLite\nanoFramework.M5AtomLite.nfproj", "{C20734A9-C944-4B2A-9CEA-A9B3F855B132}"
+Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.AtomLite", "nanoFramework.AtomLite\nanoFramework.AtomLite.nfproj", "{C20734A9-C944-4B2A-9CEA-A9B3F855B132}"
 EndProject
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "M5AtomLiteTestApp", "Tests\M5AtomLiteTestApp\M5AtomLiteTestApp.nfproj", "{7533C164-9D3E-461B-BEED-888C91AC640B}"
+Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "AtomLiteTestApp", "Tests\AtomLiteTestApp\AtomLiteTestApp.nfproj", "{7533C164-9D3E-461B-BEED-888C91AC640B}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
